google/urara: retrieve board ID from a CBFS file
Concerto board does not have the means of detecting the identity of
the device it is controlling. But it is very beneficial to be able to
use the same firmware image on Concerto boards running different
devices.

The suggested solution is to keep the device identity as a string in a
raw CBFS file called 'board_id'.

With this patch coreboot maintains a table of possible board name
strings and their matching board IDs.

BRANCH=none
BUG=chrome-os-partner:37593
TEST=verified that without the board id file addition the default
    Board ID of zero is used. Adding the file as follows:

      echo -n 'concerto' > /tmp/bid
      cbfstool /build/urara/firmware/image.serial.bin  add -f /tmp/bid  \
           -t raw -n 'board_id'

    results in firmware reporting board ID setting of 1.

board_id: failed to locate CBFS file board_id
board_id: name urara, ID 0

Change-Id: I5a02192740dc94b1ea8090092cc325fe0ac42aa6
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: f41f9b07f155f0c719c36e0cd93081205624557e
Original-Change-Id: I8341782005b101be78f5c9a6b375c8f73179c1ad
Original-Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/257170
Reviewed-on: http://review.coreboot.org/9856
Tested-by: build bot (Jenkins)
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>